backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
markdown==3.5.2
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==41.0.7
pydantic==2.5.3
pydantic-settings==2.1.0
